placed the `RouteSpec` in `specs.py` from user_specs.py
