+++ backend/tests/compare/dataset_specs.py
@@ -15,7 +15,7 @@ from typing import Callable, Optional
 
 from tests.helpers.recorder import RequestSpec
 from tests.helpers.seed import unique_suffix, TEST_PREFIX
-from tests.compare.user_specs import RouteSpec
+from tests.helpers.specs import RouteSpec
 
 # ---------------------------------------------------------------------------
 # Setup: look up test dataset + export IDs by name
