+++ backend/tests/compare/mia_specs.py
@@ -15,7 +15,7 @@ from __future__ import annotations
 import pytest
 
 from tests.helpers.recorder import RequestSpec
-from tests.compare.user_specs import RouteSpec
+from tests.helpers.specs import RouteSpec
 
 
 def _get_test_mia_annotask_id(dbm):
