+++ backend/tests/compare/sia_specs.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 import pytest
 
 from tests.helpers.recorder import RequestSpec
-from tests.compare.user_specs import RouteSpec
+from tests.helpers.specs import RouteSpec
 
 
 # ---------------------------------------------------------------------------
