+++ backend/tests/compare/test_annotask_compare.py
@@ -14,7 +14,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.annotask_specs import get_active_annotask_specs, RouteSpec
+from tests.compare.annotask_specs import get_active_annotask_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
