+++ backend/tests/compare/test_dataset_compare.py
@@ -15,7 +15,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.dataset_specs import get_active_dataset_specs, RouteSpec
+from tests.compare.dataset_specs import get_active_dataset_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
