+++ backend/tests/compare/test_filebrowser_compare.py
@@ -11,7 +11,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.filebrowser_specs import get_active_filebrowser_specs, RouteSpec
+from tests.compare.filebrowser_specs import get_active_filebrowser_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
