+++ backend/tests/compare/test_instruction_compare.py
@@ -13,7 +13,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.instruction_specs import get_active_instruction_specs, RouteSpec
+from tests.compare.instruction_specs import get_active_instruction_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute_path(path: str, context: dict) -> str:
