+++ backend/tests/compare/test_label_compare.py
@@ -15,7 +15,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.label_specs import get_active_label_specs, RouteSpec
+from tests.compare.label_specs import get_active_label_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
