+++ backend/tests/compare/test_pipeline_compare.py
@@ -12,7 +12,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.pipeline_specs import get_active_pipeline_specs, RouteSpec
+from tests.compare.pipeline_specs import get_active_pipeline_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
