+++ backend/tests/compare/test_sia_compare.py
@@ -13,7 +13,8 @@ import pytest
 
 from tests.helpers.recorder import RequestSpec, capture, save
 from tests.helpers.comparator import load_golden, assert_equal
-from tests.compare.sia_specs import get_active_sia_specs, RouteSpec
+from tests.compare.sia_specs import get_active_sia_specs
+from tests.helpers.specs import RouteSpec
 
 
 def _substitute(value, context: dict):
