+++ backend/tests/compare/user_specs.py
@@ -22,42 +22,11 @@ Routes (from lost/api/user/endpoint.py):
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
 from typing import Callable, Optional
 
 from tests.helpers.recorder import RequestSpec
+from tests.helpers.specs import RouteSpec
 from tests.helpers.seed import user_create_payload, create_test_user, cleanup_test_user, unique_suffix
-
-
-# ---------------------------------------------------------------------------
-# TestSpec — a RequestSpec + metadata about how to test it
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class RouteSpec:
-    """A test case spec: one route + how to verify it.
-
-    Attributes:
-        name: Short identifier (used in snapshot filename + test name).
-        request: The primary RequestSpec to capture.
-        skip: If True, this route is skipped (non-deterministic, can't snapshot).
-        skip_reason: Why it's skipped (for documentation).
-        follow_up: Optional follow-up RequestSpec (for mutate-then-GET).
-        setup: Optional callable(dbm) → dict to run before the primary request.
-               Returns a dict of values to substitute into the request path/body
-               (e.g. {"user_id": 42} for /api/user/<id>).
-        cleanup: Optional callable(dbm, context) to run after the test.
-    """
-
-    name: str
-    request: RequestSpec
-    skip: bool = False
-    skip_reason: str = ""
-    follow_up: Optional[RequestSpec] = None
-    setup: Optional[Callable] = None
-    cleanup: Optional[Callable] = None
-    target: str = "flask"  # "flask" or "fastapi" — which app to test against
 
 
 # ---------------------------------------------------------------------------
